git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1001 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/util/multi_callback_x.tmpl.h
@@ -164,7 +164,7 @@ struct MultiCallback$x
 			callbacks_[i]($(iP$x)$);
 	}
 	/** retrieve a callback */
-	const TCallback& operator[](int index) const
+	const TCallback& operator[](size_t index) const
 	{
 		if (index>=0 && index<=size())
 			return callbacks_[index];
